refactor(wasi): move process stream lowering to node API module (#1766)
Move the WASI process stdin/stdout/stderr call lowering out of the generic call-expression compiler and into a dedicated node-process-api module. The #1766 regression now imports process from node:process, while WASI skips registering a JS-host node:process import for that compile-time API surface.

Add #1772 to the backlog to spike edge.js as a separate Node API / WASI shim layer.
